Update dependencies and enhance API security measures
- Upgrade several Go module dependencies to their latest versions, including `golang.org/x/mod`, `golang.org/x/net`, `golang.org/x/sync`, `golang.org/x/term`, `golang.org/x/text`, and `golang.org/x/sys`, ensuring compatibility and security improvements.
- Introduce new tests to validate security measures against credential exfiltration and header injection, enhancing the robustness of API interactions.
- Implement domain normalization to ensure consistent handling of domain arguments, particularly for internationalized domain names (IDNs), improving reliability in API responses.

+++ cmd/order/order.go
@@ -293,7 +293,7 @@ func runRefund(cmd *cobra.Command, _ []string) error {
 // placed in non-USD currencies ('USD', 'CNY'), so a bare "$" would misreport
 // them. USD keeps the familiar symbol; anything else is suffixed with its code
 // rather than guessing at a symbol we may not have.
-func formatAmount(amount float32, currency *string) string {
+func formatAmount(amount float64, currency *string) string {
 	if currency == nil || *currency == "" || strings.EqualFold(*currency, "USD") {
 		return fmt.Sprintf("$%.2f", amount)
 	}
